+++ shellfoundry/utilities/standards/standards_retriever.py
@@ -1,8 +1,13 @@
-from ..cloudshell_api import create_cloudshell_client
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
 import json
+from ..cloudshell_api import create_cloudshell_client
+from shellfoundry.decorators.standards import standard_transformation
 
 
 class Standards(object):
+    @standard_transformation
     def fetch(self, **kwargs):
         alternative = kwargs.get('alternative', None)
         if not alternative: