+++ shellfoundry/utilities/template_retriever.py
@@ -1,13 +1,23 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+import click
+import os
 import requests
 import yaml
-from collections import OrderedDict
+import json
+import re
 
-from shellfoundry.models.shell_template import ShellTemplate
+from collections import OrderedDict, defaultdict
+from pkg_resources import parse_version
+
 from .filters import CompositeFilter
-from .standards import trim_standard, STANDARD_NAME_KEY
+from shellfoundry.models.shell_template import ShellTemplate
 from shellfoundry.utilities import GEN_TWO, SEPARATOR
+from shellfoundry.utilities.constants import TEMPLATE_INFO_FILE
 
 TEMPLATES_YML = 'https://raw.github.com/QualiSystems/shellfoundry/master/templates_v1.yml'
+SERVER_VERSION_KEY = "server_version"
 
 
 class TemplateRetriever(object):
@@ -17,26 +27,36 @@ class TemplateRetriever(object):
         """
 
         alternative_path = kwargs.get('alternative', None)
-        standards = kwargs.get('standards', [])
+        template_location = kwargs.get('template_location', None)
+        standards = kwargs.get('standards', {})
 
         if alternative_path:
             response = self._get_templates_from_path(alternative_path)
+            config = yaml.load(response)
+        elif template_location:
+            config = self._get_local_templates(template_location=template_location)
         else:
             response = self._get_templates_from_github()
+            config = yaml.load(response)
 
-        config = yaml.load(response)
         if not config or 'templates' not in config:
             return {}
 
-        templatesdic = OrderedDict()
+        templatesdic = defaultdict(list)
         for template in config['templates']:
-            templatesdic[template['name']] = ShellTemplate(
-                template['name'],
-                template['description'],
-                template['repository'],
-                template['min_cs_ver'],
-                self._get_standard_out_of_name(template['name']),
-                template['params'])
+
+            if template["repository"]:  # Online templates
+                standard_version = {}
+            else:
+                standard_version = template["standard_version"]
+
+            templatesdic[template["name"]].append(ShellTemplate(name=template['name'],
+                                                                description=template['description'],
+                                                                repository=template['repository'],
+                                                                min_cs_ver=template['min_cs_ver'],
+                                                                standard=self._get_standard_out_of_name(template['name']),
+                                                                standard_version=standard_version,
+                                                                params=template['params']))
 
         return self._filter_by_standards(templatesdic, standards)
 
@@ -52,12 +72,59 @@ class TemplateRetriever(object):
             response = stream.read()
         return response
 
+    def _get_local_templates(self, template_location):
+        """  """
+
+        if not template_location or not os.path.exists(template_location):
+            raise click.ClickException("Local template location empty or doesn't exist")
+        else:
+            templ_info = []
+            for root, directories, filenames in os.walk(template_location):
+                for filename in filenames:
+                    if filename == TEMPLATE_INFO_FILE:
+                        full_path = os.path.join(root, filename)
+                        standard_version = self._get_standard_version_from_template(root)
+                        with open(full_path, mode='r') as f:
+                            templ_data = json.load(f)
+                        templ_info.append({"name": templ_data.get("template_name", "Undefined"),
+                                           "description": templ_data.get("template_descr", "Undefined"),
+                                           "min_cs_ver": templ_data.get(SERVER_VERSION_KEY, "Undefined"),
+                                           "repository": "",
+                                           "standard_version": {standard_version: {"repo": root,
+                                                                                   "min_cs_ver": templ_data.get(
+                                                                                       SERVER_VERSION_KEY, "Undefined")}},
+                                           "params": {"project_name": templ_data.get("project_name", None),
+                                                      "family_name": templ_data.get("family_name", None)}})
+
+            if templ_info:
+                templates = {"templates": sorted(templ_info, key=lambda data: data["standard_version"].keys()[0])}
+            else:
+                templates = None
+
+        return templates
+
+    @staticmethod
+    def _get_standard_version_from_template(template_location):
+        """  """
+
+        for root, directories, filenames in os.walk(template_location):
+            for filename in filenames:
+                if filename == "shell-definition.yaml":
+                    with open(os.path.join(root, "shell-definition.yaml")) as stream:
+                        match = re.search(
+                            r"cloudshell_standard:\s*cloudshell_(?P<name>\S+)_standard_(?P<version>\S+)\.\w+$",
+                            stream.read(),
+                            re.MULTILINE)
+                        if match:
+                            return str(match.groupdict()["version"].replace("_", "."))
+
     @staticmethod
     def _get_standard_out_of_name(template_name, default=None):
         """
         :type template_name str
         :return:
         """
+
         type_index = 0
         standard_index = 1
         template = template_name.split(SEPARATOR)
@@ -68,17 +135,48 @@ class TemplateRetriever(object):
     @staticmethod
     def _filter_by_standards(templates, standards):
         """
-        :type templates collections.OrderedDict
-        :type standards list
+        :type templates collections.defaultdict(list)
+        :type standards dict
         :return:
         """
+
         if not standards:
-            return templates
+            return OrderedDict(sorted(templates.iteritems()))
 
-        trimmed_standards = [trim_standard(standard[STANDARD_NAME_KEY]) for standard in standards]
-        template_names = [x.name for x in templates.itervalues() if not x.standard or x.standard in trimmed_standards]  # creates a list of all matching templates names by available standard
+        filtered_templates = defaultdict(list)
+        for template_name, templates_list in templates.iteritems():
+            clear_template_name = TemplateRetriever._get_standard_out_of_name(template_name)
+            if clear_template_name is None:
+                for template in templates_list:
+                    filtered_templates[template_name].append(template)
+            elif clear_template_name in standards.keys():
+                for template in templates_list:
+                    if not template.standard_version or template.standard_version.keys()[0] in standards[clear_template_name]:
+                        if template.repository:
+                            template.min_cs_ver = TemplateRetriever._get_min_cs_version(repository=template.repository,
+                                                                                        standard_name=template.standard,
+                                                                                        standards=standards) or template.min_cs_ver
+                        filtered_templates[template_name].append(template)
 
-        return OrderedDict((name, templates[name]) for name in template_names)
+        return OrderedDict(sorted(filtered_templates.iteritems()))
+
+    @staticmethod
+    def _get_min_cs_version(repository, standard_name, standards):
+        """ Get minimal CloudShell Server Version available for provided template """
+
+        min_standard_version = unicode(min(map(parse_version, standards[standard_name])))
+        repository = repository.replace("https://github.com", "https://raw.github.com")
+        url = "/".join([repository, min_standard_version, "cookiecutter.json"])
+
+        session = requests.Session()
+        session.mount('https://', requests.adapters.HTTPAdapter(max_retries=5))
+        responce = session.get(url)
+
+        if responce.status_code == 200:
+            data = json.loads(responce.text)
+            return data.get(SERVER_VERSION_KEY, None)
+        else:
+            return
 
 
 class FilteredTemplateRetriever(object):